+++ src/main/java/io/supertokens/webserver/api/session/SessionAPI.java
@@ -123,6 +123,10 @@ public class SessionAPI extends WebserverAPI {
 
             JsonObject result = sessionInfo.toJsonObject();
 
+            if (getVersionFromRequest(req).lesserThan(SemVer.v3_0)) {
+                result.get("session").getAsJsonObject().remove("tenantId");
+            }
+
             result.addProperty("status", "OK");
 
             if (super.getVersionFromRequest(req).greaterThanOrEqualTo(SemVer.v2_21)) {
@@ -163,6 +167,10 @@ public class SessionAPI extends WebserverAPI {
             result.add("userDataInDatabase", Utils.toJsonTreeWithNulls(sessionInfo.userDataInDatabase));
 
             result.addProperty("status", "OK");
+
+            if (getVersionFromRequest(req).greaterThanOrEqualTo(SemVer.v3_0)) {
+                result.addProperty("tenantId", tenantIdentifierWithStorage.getTenantId());
+            }
 
             super.sendJsonResponse(200, result, resp);